// packages/client/src/app/routes/route-data-key.enum.ts
export enum RouteDataKey {
  TRANSLATIONS = 'key-value',
  SEO_META = 'seo-meta',
  OVERVIEW_DATA = 'overviewData',
  CONTEXT_DATA = 'contextData',
  DETAIL_DATA = 'detailData',
  WEEKLY_REPORT_LIST = 'weeklyReportWeekList',
  GEO_JSON = 'geoJson',
  GEO_JSON_2 = 'geoJson2',
  QUERY_PARAMS_MAPPING = 'queryParamsMapping',
  QUERY_PARAMS_TO_REMOVE = 'queryParamsToRemove',
  SIMPLE_GDI_OBJECT = 'simpleGdiObject',
}